Separate Schema and runtime validations into separate files
